+++ react-spring-experiment/src/pages/Pokedex/index.tsx
@@ -1,43 +1,17 @@
-import React, { useEffect, useState } from 'react';
-
-import api from '../../services/api';
-import { captalizeFirstLetter, formatId } from '../../utils';
+import React from 'react';
 
 import Card from '../../components/Card';
+import { IPokemon } from '../../App';
+
 import { Container, PokemonGrid } from './styles';
 
-export interface IPokemon {
-  id: string;
-  name: string;
-  image: string;
+interface PokedexProps {
+  pokemons: IPokemon[];
 }
 
-const Pokedex: React.FC = () => {
-  const [pokemons, setPokemons] = useState<IPokemon[]>([]);
-
-  useEffect(() => {
-    api.get('pokedex/2/').then((response) => {
-      const { data } = response;
-      const formattedData = data.pokemon_entries
-        .slice(0, 6)
-        .map(
-          (pokemon: {
-            entry_number: number;
-            pokemon_species: { name: string };
-          }) => ({
-            id: formatId(pokemon.entry_number),
-            name: captalizeFirstLetter(pokemon.pokemon_species.name),
-            image: `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/${pokemon.entry_number}.png`,
-          })
-        );
-
-      setPokemons(formattedData);
-    });
-  }, []);
-
+const Pokedex: React.FC<PokedexProps> = ({ pokemons }) => {
   return (
     <Container>
-      {pokemons.length ? <p>Tem pokemon</p> : <p>Não tem pokemon</p>}
       <PokemonGrid>
         {pokemons.map((pokemon) => (
           <Card pokemonData={pokemon} key={pokemon.id} />
