--- react-spring-experiment/src/pages/Test/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Test: React.FC = () => {
-  return <h1>Test</h1>;
-};
-
-export default Test;
